+++ SIL.Media/FFmpegRunner.cs
@@ -2,45 +2,52 @@ using System;
 using System.Collections.Generic;
 using System.Diagnostics;
 using System.IO;
+using JetBrains.Annotations;
 using SIL.CommandLineProcessing;
 using SIL.IO;
 using SIL.PlatformUtilities;
 using SIL.Progress;
+using static System.String;
 
 namespace SIL.Media
 {
 	///<summary>
-	/// FFmpeg is an open source media processing commandline library
+	/// FFMpeg is an open source media processing commandline library. Note that there is
+	/// a nuget package called FFMpegCore that wraps the exe to provide this functionality.
+	/// It has support for asynchronous processing, and with some poking around we could
+	/// maybe figure out how to use that to provide progress reporting, but it's not clear
+	/// that it would be worth it.
 	///</summary>
 	public class FFmpegRunner
 	{
+		private const string kFFMpegExe = "ffmpeg.exe";
+		private const string mp3LameCodecArg = "-acodec libmp3lame";
+
 		/// <summary>
-		/// If your app knows where FFMPEG lives, you can tell us before making any calls.
+		/// If your app knows where FFMpeg lives, you can tell us before making any calls.
 		/// </summary>
 		public static string FFmpegLocation;
 
 		/// <summary>
-		/// Find the path to ffmpeg, and remember it (some apps (like SayMore) call ffmpeg a lot)
+		/// Find the path to FFMpeg, and remember it (some apps (like SayMore) call FFMpeg a lot)
 		/// </summary>
 		/// <returns></returns>
-		static internal string LocateAndRememberFFmpeg()
+		internal static string LocateAndRememberFFMpeg()
 		{
 			if (null != FFmpegLocation) //NO! string.empty means we looked and didn't find: string.IsNullOrEmpty(s_ffmpegLocation))
 				return FFmpegLocation;
-			FFmpegLocation = LocateFFmpeg();
+			FFmpegLocation = LocateFFMpeg();
 			return FFmpegLocation;
 		}
 
 		/// <summary>
-		/// ffmpeg is more of a "compile it yourself" thing, and yet
-		/// SIL doesn't necessarily want to be redistributing something
-		/// which may violate software patents (e.g. mp3) in certain countries, so
-		/// we ask users to get it themselves.
-		/// See: http://www.ffmpeg.org/legal.html
-		/// This tries to find where they put it.
+		/// FFMpeg will typically be distributed with SIL software on Windows or automatically
+		/// installed via package dependencies on other platforms, but if something wants to
+		/// use this library and work with a version of it that the user downloaded or compiled
+		/// locally, this tries to find where they put it.
 		/// </summary>
 		/// <returns>the path, if found, else null</returns>
-		static private string LocateFFmpeg()
+		private static string LocateFFMpeg()
 		{
 			if (Platform.IsLinux)
 			{
@@ -53,18 +60,22 @@ namespace SIL.Media
 				return null;
 			}
 
-			string withApplicationDirectory = GetPathToBundledFFmpeg();
+			string withApplicationDirectory = GetPathToBundledFFMpeg();
 
-			if (!string.IsNullOrEmpty(withApplicationDirectory) && File.Exists(withApplicationDirectory))
+			if (!IsNullOrEmpty(withApplicationDirectory) && File.Exists(withApplicationDirectory))
 				return withApplicationDirectory;
+
+			var fromChoco = MediaInfo.GetFFMpegFolderFromChocoInstall(kFFMpegExe);
+			if (fromChoco != null)
+				return Path.Combine(fromChoco, kFFMpegExe);
 
 			var progFileDirs = new List<string> {
 				Environment.GetFolderPath(Environment.SpecialFolder.ProgramFilesX86),
 				Environment.GetFolderPath(Environment.SpecialFolder.ProgramFiles)
 			};
 
-			/* We DON't SUPPORT THIS ONE (it lacks some information on the output, at least as of
-			 * Julu 2010)
+			/* We DON'T SUPPORT THIS ONE (it lacks some information on the output, at least as of
+			 * July 2010)
 			 //from http://www.arachneweb.co.uk/software/windows/avchdview/ffmpeg.html
 			foreach (var path in progFileDirs)
 			{
@@ -74,50 +85,47 @@ namespace SIL.Media
 			}
 			 */
 
+			// REVIEW: I just followed the instructions in the current version of Audacity for
+			// installing FFMpeg for Audacity and the result is a folder that does not contain
+			// ffmpeg.exe. This maybe used to work, but I don't think we'll ever find ffmpeg this
+			// way now.
 			//http://manual.audacityteam.org/index.php?title=FAQ:Installation_and_Plug-Ins#installffmpeg
 			foreach (var path in progFileDirs)
 			{
-				var exePath = (Path.Combine(path, "FFmpeg for Audacity/ffmpeg.exe"));
+				var exePath = (Path.Combine(path, "FFmpeg for Audacity", kFFMpegExe));
 				if (File.Exists(exePath))
 					return exePath;
 			}
-			return string.Empty;
+			return Empty;
 		}
 
-		private static string GetPathToBundledFFmpeg()
+		private static string GetPathToBundledFFMpeg()
 		{
 			try
 			{
-				return FileLocationUtilities.GetFileDistributedWithApplication("ffmpeg", "ffmpeg.exe");
+				return FileLocationUtilities.GetFileDistributedWithApplication("ffmpeg", kFFMpegExe);
 			}
 			catch (Exception)
 			{
-				return string.Empty;
-			}
-		}
-
-
-		///<summary>
-		/// Returns false if it can't find ffmpeg
-		///</summary>
-		static public bool HaveNecessaryComponents
-		{
-			get
-			{
-				return !string.IsNullOrEmpty(LocateFFmpeg());
+				return Empty;
 			}
 		}
 
 		///<summary>
 		/// Returns false if it can't find ffmpeg
 		///</summary>
-		static private bool HaveValidFFMpegOnPath
+		public static bool HaveNecessaryComponents => !IsNullOrEmpty(LocateFFMpeg());
+
+		///<summary>
+		/// Returns false if it can't find ffmpeg
+		///</summary>
+		private static bool HaveValidFFMpegOnPath
 		{
 			get
 			{
 				if (Platform.IsWindows)
 				{
-					if (!string.IsNullOrEmpty(LocateFFmpeg()))
+					if (!IsNullOrEmpty(LocateFFMpeg()))
 						return true;
 				}
 
@@ -142,7 +150,7 @@ namespace SIL.Media
 
 		/// <summary>
 		/// Extracts the audio from a video. Note, it will fail if the file exists, so the client
-		/// is resonsible for verifying with the user and deleting the file before calling this.
+		/// is responsible for verifying with the user and deleting the file before calling this.
 		/// </summary>
 		/// <param name="inputPath"></param>
 		/// <param name="outputPath"></param>
@@ -151,18 +159,11 @@ namespace SIL.Media
 		/// <returns>log of the run</returns>
 		public static ExecutionResult ExtractMp3Audio(string inputPath, string outputPath, int channels, IProgress progress)
 		{
-			if (string.IsNullOrEmpty(LocateFFmpeg()))
-			{
-				return new ExecutionResult() { StandardError = "Could not locate FFMpeg" };
-			}
+			if (IsNullOrEmpty(LocateFFMpeg()))
+				return new ExecutionResult { StandardError = "Could not locate FFMpeg" };
 
-			var arguments = string.Format("-i \"{0}\" -vn -acodec libmp3lame -ac {1} \"{2}\"", inputPath, channels, outputPath);
-			var result = CommandLineRunner.Run(LocateAndRememberFFmpeg(),
-														arguments,
-														Environment.CurrentDirectory,
-														60 * 10, //10 minutes
-														progress
-				);
+			var arguments = $"-i \"{inputPath}\" -vn {mp3LameCodecArg} -ac {channels} \"{outputPath}\"";
+			var result = RunFFMpeg(arguments, progress);
 
 			progress.WriteVerbose(result.StandardOutput);
 
@@ -174,7 +175,7 @@ namespace SIL.Media
 				{
 					ExitCode = 0,
 					StandardOutput = result.StandardOutput,
-					StandardError = string.Empty
+					StandardError = Empty
 				};
 				return doctoredResult;
 			}
@@ -183,9 +184,10 @@ namespace SIL.Media
 
 			return result;
 		}
+
 		/// <summary>
 		/// Extracts the audio from a video. Note, it will fail if the file exists, so the client
-		/// is resonsible for verifying with the user and deleting the file before calling this.
+		/// is responsible for verifying with the user and deleting the file before calling this.
 		/// </summary>
 		/// <param name="inputPath"></param>
 		/// <param name="outputPath"></param>
@@ -194,19 +196,12 @@ namespace SIL.Media
 		/// <returns>log of the run</returns>
 		public static ExecutionResult ExtractOggAudio(string inputPath, string outputPath, int channels, IProgress progress)
 		{
-			if (string.IsNullOrEmpty(LocateFFmpeg()))
-			{
-				return new ExecutionResult() { StandardError = "Could not locate FFMpeg" };
-			}
+			if (IsNullOrEmpty(LocateFFMpeg()))
+				return new ExecutionResult { StandardError = "Could not locate FFMpeg" };
 
-			var arguments = string.Format("-i \"{0}\" -vn -acodec vorbis -ac {1} \"{2}\"", inputPath, channels, outputPath);
+			var arguments = $"-i \"{inputPath}\" -vn -acodec vorbis -ac {channels} \"{outputPath}\"";
 			progress.WriteMessage("ffmpeg " + arguments);
-			var result = CommandLineRunner.Run(LocateAndRememberFFmpeg(),
-														arguments,
-														Environment.CurrentDirectory,
-														60 * 10, //10 minutes
-														progress
-				);
+			var result = RunFFMpeg(arguments, progress);
 
 			progress.WriteVerbose(result.StandardOutput);
 
@@ -218,7 +213,7 @@ namespace SIL.Media
 				{
 					ExitCode = 0,
 					StandardOutput = result.StandardOutput,
-					StandardError = string.Empty
+					StandardError = Empty
 				};
 				return doctoredResult;
 			}
@@ -227,7 +222,6 @@ namespace SIL.Media
 
 			return result;
 		}
-
 
 		/// <summary>
 		/// Extracts the audio from a video. Note, it will fail if the file exists, so the client
@@ -238,6 +232,7 @@ namespace SIL.Media
 		/// <param name="channels">0 for same, 1 for mono, 2 for stereo</param>
 		/// <param name="progress"></param>
 		/// <returns>log of the run</returns>
+		[PublicAPI]
 		public static ExecutionResult ExtractBestQualityWavAudio(string inputPath, string outputPath, int channels, IProgress progress)
 		{
 			return ExtractAudio(inputPath, outputPath, "copy", 0, channels, progress);
@@ -254,6 +249,7 @@ namespace SIL.Media
 		/// <param name="channels">0 for same, 1 for mono, 2 for stereo</param>
 		/// <param name="progress"></param>
 		/// <returns>log of the run</returns>
+		[PublicAPI]
 		public static ExecutionResult ExtractPcmAudio(string inputPath, string outputPath,
 			int bitsPerSample, int sampleRate, int channels, IProgress progress)
 		{
@@ -284,30 +280,23 @@ namespace SIL.Media
 		private static ExecutionResult ExtractAudio(string inputPath, string outputPath,
 			string audioCodec, int sampleRate, int channels, IProgress progress)
 		{
-			if (string.IsNullOrEmpty(LocateFFmpeg()))
-			{
-				return new ExecutionResult() { StandardError = "Could not locate FFMpeg" };
-			}
+			if (IsNullOrEmpty(LocateFFMpeg()))
+				return new ExecutionResult { StandardError = "Could not locate FFMpeg" };
 
 			var sampleRateArg = "";
 			if (sampleRate > 0)
-				sampleRateArg = string.Format("-ar {0}", sampleRate);
+				sampleRateArg = $"-ar {sampleRate}";
 
 			//TODO: this will output whatever mp3 or wav or whatever is in the video... might not be wav at all!
 			var channelsArg = "";
 			if (channels > 0)
-				channelsArg = string.Format(" -ac {0}", channels);
+				channelsArg = $" -ac {channels}";
 
-			var arguments = string.Format("-i \"{0}\" -vn -acodec {1}  {2} {3} \"{4}\"",
-				inputPath, audioCodec, sampleRateArg, channelsArg, outputPath);
+			var arguments = $"-i \"{inputPath}\" -vn -acodec {audioCodec} {sampleRateArg} {channelsArg} \"{outputPath}\"";
 
 			progress.WriteMessage("ffmpeg " + arguments);
 
-			var result = CommandLineRunner.Run(LocateAndRememberFFmpeg(),
-														arguments,
-														Environment.CurrentDirectory,
-														60 * 10, //10 minutes
-														progress);
+			var result = RunFFMpeg(arguments, progress);
 
 			progress.WriteVerbose(result.StandardOutput);
 
@@ -319,7 +308,7 @@ namespace SIL.Media
 				{
 					ExitCode = 0,
 					StandardOutput = result.StandardOutput,
-					StandardError = string.Empty
+					StandardError = Empty
 				};
 				return doctoredResult;
 			}
@@ -338,17 +327,13 @@ namespace SIL.Media
 		public static ExecutionResult ChangeNumberOfAudioChannels(string inputPath,
 			string outputPath, int channels, IProgress progress)
 		{
-			if (string.IsNullOrEmpty(LocateFFmpeg()))
+			if (IsNullOrEmpty(LocateFFMpeg()))
 				return new ExecutionResult { StandardError = "Could not locate FFMpeg" };
 
-			var arguments = string.Format("-i \"{0}\" -vn -ac {1} \"{2}\"",
+			var arguments = Format("-i \"{0}\" -vn -ac {1} \"{2}\"",
 				inputPath, channels, outputPath);
 
-			var result = CommandLineRunner.Run(LocateAndRememberFFmpeg(),
-							arguments,
-							Environment.CurrentDirectory,
-							60 * 10, //10 minutes
-							progress);
+			var result = RunFFMpeg(arguments, progress);
 
 			progress.WriteVerbose(result.StandardOutput);
 
@@ -359,7 +344,7 @@ namespace SIL.Media
 				{
 					ExitCode = 0,
 					StandardOutput = result.StandardOutput,
-					StandardError = string.Empty
+					StandardError = Empty
 				};
 
 				return doctoredResult;
@@ -378,23 +363,14 @@ namespace SIL.Media
 		/// <returns>log of the run</returns>
 		public static ExecutionResult MakeLowQualityCompressedAudio(string inputPath, string outputPath, IProgress progress)
 		{
-			if (string.IsNullOrEmpty(LocateAndRememberFFmpeg()))
-			{
-				return new ExecutionResult() { StandardError = "Could not locate FFMpeg" };
-			}
+			if (IsNullOrEmpty(LocateAndRememberFFMpeg()))
+				return new ExecutionResult { StandardError = "Could not locate FFMpeg" };
 
-			var arguments = "-i \"" + inputPath + "\" -acodec libmp3lame -ac 1 -ar 8000 \"" + outputPath + "\"";
-
+			var arguments = $"-i \"{inputPath}\" {mp3LameCodecArg} -ac 1 -ar 8000 \"{outputPath}\"";
 
 			progress.WriteMessage("ffmpeg " + arguments);
 
-
-			var result = CommandLineRunner.Run(LocateAndRememberFFmpeg(),
-														arguments,
-														Environment.CurrentDirectory,
-														60 * 10, //10 minutes
-														progress
-				);
+			var result = RunFFMpeg(arguments, progress);
 
 			progress.WriteVerbose(result.StandardOutput);
 
@@ -407,7 +383,7 @@ namespace SIL.Media
 				{
 					ExitCode = 0,
 					StandardOutput = result.StandardOutput,
-					StandardError = string.Empty
+					StandardError = Empty
 				};
 			}
 			if (result.StandardError.ToLower().Contains("error")
@@ -430,29 +406,21 @@ namespace SIL.Media
 		/// <returns>log of the run</returns>
 		public static ExecutionResult MakeLowQualitySmallVideo(string inputPath, string outputPath, int maxSeconds, IProgress progress)
 		{
-			if (string.IsNullOrEmpty(LocateAndRememberFFmpeg()))
-			{
-				return new ExecutionResult() { StandardError = "Could not locate FFMpeg" };
-			}
+			if (IsNullOrEmpty(LocateAndRememberFFMpeg()))
+				return new ExecutionResult { StandardError = "Could not locate FFMpeg" };
 
 			// isn't working: var arguments = "-i \"" + inputPath + "\" -vcodec mpeg4 -s 160x120 -b 800  -acodec libmp3lame -ar 22050 -ab 32k -ac 1 \"" + outputPath + "\"";
-			var arguments = "-i \"" + inputPath +
-							"\" -vcodec mpeg4 -s 160x120 -b 800 -acodec libmp3lame -ar 22050 -ab 32k -ac 1 ";
+			var arguments = $"-i \"{inputPath}\" -vcodec mpeg4 -s 160x120 -b 800 " +
+				$"{mp3LameCodecArg} -ar 22050 -ab 32k -ac 1 ";
 			if (maxSeconds > 0)
-				arguments += " -t " + maxSeconds + " ";
-			arguments += "\"" + outputPath + "\"";
+				arguments += $" -t {maxSeconds} ";
+			arguments += $"\"{outputPath}\"";
 
 			progress.WriteMessage("ffmpeg " + arguments);
 
-			var result = CommandLineRunner.Run(LocateAndRememberFFmpeg(),
-														arguments,
-														Environment.CurrentDirectory,
-														60 * 10, //10 minutes
-														progress
-				);
+			var result = RunFFMpeg(arguments, progress);
 
 			progress.WriteVerbose(result.StandardOutput);
-
 
 			//hide a meaningless error produced by some versions of liblame
 			if (result.StandardError.Contains("lame: output buffer too small")
@@ -462,9 +430,8 @@ namespace SIL.Media
 				{
 					ExitCode = 0,
 					StandardOutput = result.StandardOutput,
-					StandardError = string.Empty
+					StandardError = Empty
 				};
-
 			}
 			if (result.StandardError.ToLower().Contains("error") //ffmpeg always outputs config info to standarderror
 				|| result.StandardError.ToLower().Contains("unable to")
@@ -479,31 +446,32 @@ namespace SIL.Media
 		/// Converts to low-quality, small picture
 		/// </summary>
 		/// <returns>log of the run</returns>
+		[PublicAPI]
 		public static ExecutionResult MakeLowQualitySmallPicture(string inputPath, string outputPath, IProgress progress)
 		{
-			if (string.IsNullOrEmpty(LocateAndRememberFFmpeg()))
-			{
-				return new ExecutionResult() { StandardError = "Could not locate FFMpeg" };
-			}
+			if (IsNullOrEmpty(LocateAndRememberFFMpeg()))
+				return new ExecutionResult { StandardError = "Could not locate FFMpeg" };
 
 			//enhance: how to lower the quality?
 
-			var arguments = "-i \"" + inputPath + "\" -f image2  -s 176x144 \"" + outputPath + "\"";
+			var arguments = $"-i \"{inputPath}\" -f image2  -s 176x144 \"{outputPath}\"";
 
 			progress.WriteMessage("ffmpeg " + arguments);
 
-			var result = CommandLineRunner.Run(LocateAndRememberFFmpeg(),
-														arguments,
-														Environment.CurrentDirectory,
-														60 * 10, //10 minutes
-														progress
-				);
+			var result = RunFFMpeg(arguments, progress);
 
 			progress.WriteVerbose(result.StandardOutput);
 			if (result.StandardError.ToLower().Contains("error")) //ffmpeg always outputs config info to standarderror
 				progress.WriteError(result.StandardError);
 
 			return result;
+		}
+
+		private static ExecutionResult RunFFMpeg(string arguments, IProgress progress)
+		{
+			const int timeout = 600; // 60 * 10 = 10 minutes
+			return CommandLineRunner.Run(LocateAndRememberFFMpeg(), arguments,
+				Environment.CurrentDirectory, timeout, progress);
 		}
 	}
 }