+++ SIL.Media/MediaInfo.cs
@@ -1,300 +1,257 @@
-﻿using System;
+using System;
 using System.Diagnostics;
-using System.Text.RegularExpressions;
+using System.IO;
+using FFMpegCore;
+using FFMpegCore.Exceptions;
+using SIL.PlatformUtilities;
+using SIL.Reporting;
+using static System.String;
+using static SIL.IO.FileLocationUtilities;
 
 namespace SIL.Media
 {
 	/// <summary>
-	/// This class uses ffmpeg to gather information about media streams
+	/// This class uses FFProbe (via FFMpegCore) to gather information about media streams.
 	/// </summary>
 	public class MediaInfo
 	{
+		private const string kFFProbeExe = "ffprobe.exe";
+		private static string s_ffProbeFolder;
+		
 		/// <summary>
-		/// The actual ffmpeg output
+		/// Returns false if it can't find FFProbe
 		/// </summary>
-		public string RawData;
-
-		private  MediaInfo(string ffmpegOutput)
-		{
-			RawData = ffmpegOutput;
-			Audio = new AudioInfo(ffmpegOutput);
-			if (ffmpegOutput.Contains("Video:"))
-			{
-				Video = new VideoInfo(ffmpegOutput);
-			}
-		}
-
-		///<summary>
-		/// Returns false if it can't find ffmpeg
-		///</summary>
-		static public bool HaveNecessaryComponents
-		{
-			get
-			{
-				return FFmpegRunner.HaveNecessaryComponents;
-			}
-		}
-
-		static public MediaInfo GetInfo(string path)
-		{
-			if(!HaveNecessaryComponents)
-			{
-				return new MediaInfo("Could not locate FFMpeg");
-			}
-			var p = new Process();
-			p.StartInfo.FileName = FFmpegRunner.LocateAndRememberFFmpeg();
-			p.StartInfo.Arguments = ( "-i " + "\""+path+"\"");
-			p.StartInfo.RedirectStandardError = true;
-			p.StartInfo.UseShellExecute = false;
-			p.StartInfo.CreateNoWindow = true;
-			p.Start();
-
-			string ffmpegOutput = p.StandardError.ReadToEnd();
-			if(!p.WaitForExit(3000))
-			{
-				try
-				{
-					p.Kill();
-				}
-				catch(Exception)
-				{ //couldn't stop it, oh well.
-				}
-				throw new ApplicationException("FFMpeg seems to have hung");
-			}
-			return new MediaInfo(ffmpegOutput);
-		}
-
-		public AudioInfo Audio { get; private set;}
-		public VideoInfo Video { get; private set;}
-		//future public ImageInfo Image { get; private set;}
+		public static bool HaveNecessaryComponents => !IsNullOrEmpty(LocateAndRememberFFProbe());
 
 		/// <summary>
-		/// used by both audio and video, so it lives up here
+		/// The folder where FFProbe should be found. An application can use this to set
+		/// the location before making any calls.
 		/// </summary>
-		internal static TimeSpan GetDuration(string ffmpegOutput)
+		/// <exception cref="DirectoryNotFoundException">Folder does not exist</exception>
+		/// <exception cref="FileNotFoundException">Folder does not contain FFProbe.exe</exception>
+		/// <remarks>If set to <see cref="Empty"/>, indicates to this library that
+		/// FFProbe is not installed, and therefore methods to retrieve media info will
+		/// fail unconditionally (i.e. they will throw an exception)</remarks>
+		public static string FFProbeFolder
 		{
-				var match = Regex.Match(ffmpegOutput, "Duration: ([^,]*),");
-				if (match.Groups.Count == 2)
+			get => s_ffProbeFolder;
+			set
+			{
+				if (value != Empty)
 				{
-					var durationString = match.Groups[1].Value;
+					if (!Directory.Exists(value))
+						throw new DirectoryNotFoundException("Directory not found: " + value);
 
-					TimeSpan durationTime;
-					if (TimeSpan.TryParse(durationString, out durationTime))
+					if (Platform.IsWindows && !File.Exists(Path.Combine(value, kFFProbeExe)))
 					{
-						return durationTime;
+						throw new FileNotFoundException(
+							"Path is not a folder containing FFProbe.exe: " + value);
 					}
 				}
-				return default(TimeSpan);
+
+				s_ffProbeFolder = value;
+			}
 		}
+
+		/// <summary>
+		/// Find the path to FFProbe, and remember it.
+		/// </summary>
+		/// <returns></returns>
+		private static string LocateAndRememberFFProbe()
+		{
+			if (null != FFProbeFolder)
+				return FFProbeFolder;
+			try
+			{
+				FFProbeFolder = GetPresumedFFProbeFolder();
+			}
+			catch (Exception e)
+			{
+				Console.WriteLine(e);
+				return null;
+			}
+			return FFProbeFolder;
+		}
+
+		/// <summary>
+		/// On Windows, FFProbe will typically be distributed with SIL software, but if something
+		/// wants to use this library and work with a version of it that the user downloaded or
+		/// compiled locally, this tries to find where they put it. On other platforms, ffprobe
+		/// should be installed in the standard location by the package manager.
+		/// </summary>
+		/// <returns>A folder where FFProbe can be found; else <see cref="Empty"/></returns>
+		private static string GetPresumedFFProbeFolder()
+		{
+			var folder = GlobalFFOptions.Current.BinaryFolder;
+
+			if (Platform.IsWindows)
+			{
+				if (IsNullOrEmpty(folder) || !File.Exists(Path.Combine(folder, kFFProbeExe)))
+				{
+					var withApplicationDirectory =
+						GetFileDistributedWithApplication(true, "ffmpeg", kFFProbeExe) ??
+						GetFileDistributedWithApplication(true, "ffprobe", kFFProbeExe);
+
+					folder = !IsNullOrEmpty(withApplicationDirectory) ?
+						Path.GetDirectoryName(withApplicationDirectory) :
+						GetFFMpegFolderFromChocoInstall();
+
+					if (folder != null)
+					{
+						GlobalFFOptions.Configure(new FFOptions { BinaryFolder = folder });
+					}
+					else
+						folder = Empty;
+				}
+			}
+
+			return folder;
+		}
+
+		internal static string GetFFMpegFolderFromChocoInstall(string exeNeeded = kFFProbeExe)
+		{
+			try
+			{
+				var programData = Environment.GetFolderPath(Environment.SpecialFolder
+					.CommonApplicationData);
+
+				var folder = Path.Combine(programData, "chocolatey", "lib", "ffmpeg", "tools",
+					"ffmpeg", "bin");
+				if (!File.Exists(Path.Combine(folder, exeNeeded)))
+					folder = null;
+				return folder;
+			}
+			catch (Exception e)
+			{
+				Console.WriteLine(e);
+				return null;
+			}
+		}
+
+		public IMediaAnalysis AnalysisData { get; }
+
+		private MediaInfo(IMediaAnalysis mediaAnalysis)
+		{
+			AnalysisData = mediaAnalysis;
+			if (mediaAnalysis != null)
+			{
+				if (mediaAnalysis.PrimaryAudioStream != null)
+					Audio = new AudioInfo(mediaAnalysis);
+				if (mediaAnalysis.PrimaryVideoStream != null)
+					Video = new VideoInfo(mediaAnalysis);
+			}
+		}
+
+		public static string MissingComponentMessage => Format(
+			Localizer.GetString("FFProbeMissing", "Could not locate {0}",
+				"Param is the name of a utility program required to be installed."),
+			"FFProbe");
+
+		/// <summary>
+		/// Gets the media information for the requested media file. If media information
+		/// the error is logged and an object is returned with null Audio and Video.
+		/// </summary>
+		public static MediaInfo GetInfo(string path)
+		{
+			if (!HaveNecessaryComponents)
+				throw new ApplicationException(MissingComponentMessage);
+
+			var info = GetInfo(path, out var e);
+			if (e != null)
+			{
+				Logger.WriteError(e);
+				return new MediaInfo(null);
+			}
+
+			return info;
+		}
+
+		/// <summary>
+		/// Gets the media information for the requested media file
+		/// </summary>
+		/// <param name="path">Path of the media file</param>
+		/// <param name="error">If this method returns null, this will be set to indicate the error
+		/// that occurred. If FFProbe cannot be located, this will be an
+		/// <see cref="ApplicationException"/>. Otherwise, it will probably be an
+		/// <see cref="FFMpegException"/>.</param>
+		/// <returns>A new MediaInfo object, or null if media information could not be retrieved.</returns>
+		public static MediaInfo GetInfo(string path, out Exception error)
+		{
+			try
+			{
+				error = null;
+				return new MediaInfo(FFProbe.Analyse(path));
+			}
+			catch (Exception e)
+			{
+				error = e;
+				return null;
+			}
+		}
+
+		/// <summary>
+		/// Information about the primary audio track, if any
+		/// </summary>
+		public AudioInfo Audio { get; }
+		/// <summary>
+		/// Information about the primary video track, if any
+		/// </summary>
+		public VideoInfo Video { get; }
+
+		//future public ImageInfo Image { get; private set;}
 
 		public class AudioInfo
 		{
-
-			internal AudioInfo(string ffmpegOutput)
+			internal AudioInfo(IMediaAnalysis mediaAnalysis)
 			{
-				ExtractDuration(ffmpegOutput);
-				ExtractBitDepth(ffmpegOutput);
-				ExtractSamplesPerSecond(ffmpegOutput);
-				ExtractChannels(ffmpegOutput);
-				ExtractEncoding(ffmpegOutput);
+				var primaryAudio = mediaAnalysis.PrimaryAudioStream;
+				Debug.Assert(primaryAudio != null);
+				Duration = primaryAudio.Duration;
+				Encoding = primaryAudio.CodecName;
+				ChannelCount = primaryAudio.Channels;
+				BitDepth = primaryAudio.BitDepth ?? default;
+				SamplesPerSecond = mediaAnalysis.PrimaryAudioStream?.SampleRateHz ?? default;
 			}
 
-			private void ExtractEncoding(string ffmpegOutput)
-			{
-
-				var match = Regex.Match(ffmpegOutput, "Audio: ([^,]*)");
-				if (match.Groups.Count == 2)
-				{
-					Encoding = match.Groups[1].Value;
-				}
-			}
-
-			private void ExtractChannels(string ffmpegOutput)
-			{
-				var regex = new Regex(", ([^,]*) channels");
-				if (regex.IsMatch(ffmpegOutput))
-				{
-					var match = regex.Match(ffmpegOutput);
-					var value = match.Groups[1].Value;
-
-					int channelCount;
-
-					if (int.TryParse(value, out channelCount))
-					{
-						this.ChannelCount = channelCount;
-					}
-					return;
-				}
-				// older versions (i.e. 9.18 which comes with trusty) have a slightly different
-				// output format
-				regex = new Regex(", (mono|stereo|quad|hexagonal|octagonal|downmix|[2-7].[0-1](\\(.\\))?),");
-				if (regex.IsMatch(ffmpegOutput))
-				{
-					var match = regex.Match(ffmpegOutput);
-					var value = match.Groups[1].Value;
-					switch (value)
-					{
-						case "mono":
-							this.ChannelCount = 1;
-							break;
-						case "stereo":
-						case "downmix":
-							this.ChannelCount = 2;
-							break;
-						case "2.1":
-						case "3.0":
-						case "3.0(back)":
-							this.ChannelCount = 3;
-							break;
-						case "3.1":
-						case "4.0":
-						case "quad":
-						case "quad(side)":
-							this.ChannelCount = 4;
-							break;
-						case "4.1":
-						case "5.0":
-							this.ChannelCount = 5;
-							break;
-						case "5.1":
-						case "6.0":
-						case "6.0(front)":
-						case "hexagonal":
-							this.ChannelCount = 6;
-							break;
-						case "6.1":
-						case "6.1(front)":
-						case "7.0":
-						case "7.0(front)":
-							this.ChannelCount = 7;
-							break;
-						case "7.1":
-						case "7.1(wide)":
-						case "octagonal":
-							this.ChannelCount = 8;
-							break;
-					}
-				}
-			}
-
-			private void ExtractSamplesPerSecond(string ffmpegOutput)
-			{
-				var match = Regex.Match(ffmpegOutput, ", ([^,]*) Hz");
-				if (match.Groups.Count == 2)
-				{
-					var value = match.Groups[1].Value;
-
-					int frequency;
-
-					if (int.TryParse(value, out frequency))
-					{
-						this.SamplesPerSecond = frequency;
-					}
-				}
-			}
-
-			private void ExtractBitDepth(string ffmpegOutput)
-			{
-
-				var match = Regex.Match(ffmpegOutput, @"pcm_s(\d+)");
-				if (match.Groups.Count == 2)
-				{
-					var value = match.Groups[1].Value;
-
-					//todo: find out if we will really get s16, s24, s96, etc.
-					int depth;
-					if (int.TryParse(value, out depth))
-					{
-						BitDepth = depth;
-						return;
-					}
-				}
-				//no pcm found, use the myserious "s24, s32" stuff, which may be misleading
-
-				 match = Regex.Match(ffmpegOutput, @", s(\d+),");
-				if (match.Groups.Count == 2)
-				{
-					var value = match.Groups[1].Value;
-
-					//todo: find out if we will really get s16, s24, s96, etc.
-					int depth;
-					if (int.TryParse(value, out depth))
-					{
-						BitDepth = depth;
-					}
-				}
-			}
-
-			private void ExtractDuration(string ffmpegOutput)
-			{
-				Duration = MediaInfo.GetDuration(ffmpegOutput);
-			}
-
-
-
-			public TimeSpan Duration { get; private set;}
-			public int ChannelCount { get; private set;}
-			public int SamplesPerSecond { get; private set;}
-			public int BitDepth { get; private set;}
-			public string Encoding { get; private set;}
+			/// <summary>
+			/// Note that since a media file might contain multiple audio and video tracks
+			/// and they might start and stop at different times, the total duration reported
+			/// by <see cref="MediaInfo.AnalysisData"/> could be greater than this duration.
+			/// </summary>
+			public TimeSpan Duration { get; }
+			public int ChannelCount { get; }
+			public int SamplesPerSecond { get; }
+			public int BitDepth { get; }
+			public string Encoding { get; }
 		}
-
 
 		public class VideoInfo
 		{
-			internal VideoInfo(string ffmpegOutput)
+			internal VideoInfo(IMediaAnalysis mediaAnalysis)
 			{
-				ExtractDuration(ffmpegOutput);
-				ExtractEncoding(ffmpegOutput);
-				ExtractFramesPerSecond(ffmpegOutput);
-				ExtractResolution(ffmpegOutput);
+				var primaryVideo = mediaAnalysis.PrimaryVideoStream;
+				Debug.Assert(primaryVideo != null);
+				Duration = primaryVideo.Duration;
+				Encoding = primaryVideo.CodecName;
+				FrameRate = primaryVideo.FrameRate;
+				Resolution = $"{primaryVideo.Width}x{primaryVideo.Height}";
 			}
 
-			private void ExtractResolution(string ffmpegOutput)
-			{
-				//e.g. , 176x144 [PAR 12:11 DAR 4:3],
-				//e.g. p, 320x240, 1
-				var match = Regex.Match(ffmpegOutput, @" (\d+x\d+)[, ]");//sometimes followed by space, sometimes comma
-				if (match.Groups.Count == 2)
-				{
-					Resolution = match.Groups[1].Value;
-				}
-			}
-
-			private void ExtractFramesPerSecond(string ffmpegOutput)
-			{
-				var match = Regex.Match(ffmpegOutput, ", ([^,]*) fps");
-				if (match.Groups.Count == 2)
-				{
-					int fps;
-					if (int.TryParse(match.Groups[1].Value, out fps))
-					{
-						FramesPerSecond = fps;
-					}
-				}
-			}
-
-			private void ExtractEncoding(string ffmpegOutput)
-			{
-				// DG 201303: ffmpeg gives more info like mpeg4 (Simple Profile)
-				// revert match to "Video: ([^,]*)" if that would be useful
-				var match = Regex.Match(ffmpegOutput, "Video: ([^,^(]*)");
-				if (match.Groups.Count == 2)
-				{
-					Encoding = match.Groups[1].Value.Trim();
-				}
-			}
-
-			private void ExtractDuration(string ffmpegOutput)
-			{
-				Duration = MediaInfo.GetDuration(ffmpegOutput);
-			}
-
-			public TimeSpan Duration { get; private set;}
-			public int FramesPerSecond { get; private set;}
-			public string Resolution { get; private set;}
-			public string Encoding { get; private set;}
+			/// <summary>
+			/// Note that since a media file might contain multiple audio and video tracks
+			/// and they might start and stop at different times, the total duration reported
+			/// by <see cref="MediaInfo.AnalysisData"/> could be greater than this duration.
+			/// </summary>
+			public TimeSpan Duration { get; }
+			// For backward compatibility, we want to support FramesPerSecond as an integer value,
+			// but in case clients want the real rate, we supply that as well.
+			// See https://www.hdhead.com/?p=108
+			public double FrameRate { get; }
+			public int FramesPerSecond => (int)Math.Round(FrameRate, 0);
+			public string Resolution { get; }
+			public string Encoding { get; }
 		}
+
 		/* future
 		public class ImageInfo
 		{